a rough-and-tumble demo of the charts
The code needs work!

I ended up not using graphite: it had some issues that made it hard to
control the data presentation. Raw plot was not nearly as difficult as I
expected.

Not reactive, and bakes in assumptions about the "Cabin" column and
current data. But works well enough for the demo data I have.